+++ src/QNDNSClient.h
@@ -11,8 +11,9 @@
 #include <functional>
 
 #include <IPAddress.h>
-#include <lwip/ip_addr.h>
-#include <lwip/opt.h>
+
+#include "lwip/ip_addr.h"
+#include "lwip/opt.h"
 
 namespace qindesign {
 namespace network {
@@ -21,7 +22,7 @@ namespace network {
 class DNSClient final {
  public:
   // Return the maximum number of DNS servers.
-  static int maxServers() {
+  static constexpr int maxServers() {
     return DNS_MAX_SERVERS;
   }